refactor: convertTime into util function
