+++ src/commands/core/stats.ts
@@ -3,6 +3,7 @@ import { MessageEmbed } from 'discord.js';
 import Command from '../../lib/structures/Command';
 import { TypicalMessage } from '../../lib/types/typicalbot';
 import { Modes, Links } from '../../lib/utils/constants';
+import { convertTime } from '../../lib/utils/util';
 
 export default class extends Command {
     dm = true;
@@ -22,7 +23,7 @@ export default class extends Command {
             totalRAM
         ] = await Promise.all(paths.map((path) => this.client.fetchData(path)));
 
-        const uptime = this.client.helpers.convertTime.execute(message, this.client.uptime ?? 0);
+        const uptime = convertTime(message, this.client.uptime ?? 0);
         if (!message.embeddable)
             return message.send(message.translate('core/stats:TEXT', {
                 uptime,
